requirements.txt
requests==2.32.0
beautifulsoup4==4.6.0
readability-lxml==0.8.1
lxml_html_clean
ollama>=0.1.0
scikit-learn>=1.0.0
sumy>=0.11.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0